5.1.2 proposal
These changes, if approved shouldn't be held back for donators since the change they make is so insignificant for normal use. For modding by changing the text colour of buttons, on the other hand, it is very important since this opens up more options for badges and buttons.

This code is not expected to be used in its present condition,, but rather, for demonstration purposes only and should be rejected in favour for implementing yourself. Talk to me (w/ :heart:) in a DM whether or not these changes will be integrated.